+++ configs/bair/CrevNet.py
@@ -5,4 +5,4 @@ predictor_rnn_layers = 8
 # training
 # lr = 5e-3
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'