+++ configs/bair/DMVFN.py
@@ -13,4 +13,4 @@ coef = 0.5
 # training
 # lr = 1e-4
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'