+++ configs/bair/MAU.py
@@ -17,4 +17,4 @@ model_mode = 'normal'
 # training
 # lr = 1e-3
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'