+++ configs/bair/PhyDNet.py
@@ -4,4 +4,4 @@ patch_size = 2
 # training
 # lr = 1e-3
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'