+++ configs/bair/PredNet.py
@@ -14,4 +14,4 @@ LSTM_inner_activation = 'hard_sigmoid'
 # training
 lr = 1e-3
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'