+++ configs/bair/PredRNNpp.py
@@ -18,4 +18,4 @@ layer_norm = 0
 # training
 # lr = 1e-4
 batch_size = 16
-sched = 'onecycle'
+# sched = 'onecycle'