+++ configs/bair/SimVP.py
@@ -11,5 +11,5 @@ N_S = 2
 # lr = 1e-3
 batch_size = 16
 drop_path = 0.1
-sched = 'onecycle'
+# sched = 'onecycle'
 # --broadcast_buffers is needed