+++ openstl/utils/main_utils.py
@@ -212,7 +212,7 @@ def update_config(args, config, exclude_keys=list()):
     assert isinstance(args, dict) and isinstance(config, dict)
     for k in config.keys():
         if args.get(k, False):
-            if args[k] != config[k] and k not in exclude_keys:
+            if args[k] != config[k] and k not in exclude_keys and args[k] is not None:
                 print(f'overwrite config key -- {k}: {config[k]} -> {args[k]}')
             else:
                 args[k] = config[k]